requirements.txt
accelerate==0.25.0
aiohttp==3.8.6
aiosqlite==0.17.0
alembic==1.12.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
argcomplete==3.2.0
asgiref==3.7.2
asyncio==3.4.3
atpublic==4.0
backoff==2.2.1
backports.functools-lru-cache==1.6.6
black==23.11.0
blinker==1.7.0
blis==0.7.11
boltons==23.1.1
bson==0.5.10
cachelib==0.9.0
cattrs==23.2.3
certifi==2023.7.22
cfgv==3.4.0
click-plugins==1.1.1
clickclick==20.10.2
cloudpickle==2.2.1
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contextlib2==21.6.0
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==2.0.1
cycler==0.12.1
databricks-cli==0.18.0
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
email-validator==1.3.1
execnet==2.0.2
flake8==6.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flufl.lock==8.0.2
fonttools==4.44.0
google-re2==1.1
googleapis-common-protos==1.62.0
grandalf==0.8
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
hpack==4.0.0
httpx==0.25.2
huggingface-hub==0.17.3
identify==2.5.33
inflection==0.5.1
iniconfig==2.0.0
ipython-genutils==0.2.0
iso8601==1.1.0
isodate==0.6.1
itsdangerous==2.1.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.5.1
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
merge-requirements==0.6
meteomatics==2.10.0
mkdocs==1.5.3
mlflow==2.8.0
mock==5.1.0
motor==3.3.1
mpmath==1.3.0
mypy-extensions==1.0.0
nanotime==0.5.2
nodeenv==1.8.0
numpy==1.26.2
oauthlib==3.2.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
pandas==2.1.4
path==16.9.0
path.py==12.5.0
pathtools==0.1.2
pathy==0.10.3
pendulum==2.1.2
Pillow==10.1.0
pipreqs==0.4.13
playwright==1.40.0
pre-commit==3.5.0
prettytable==3.9.0
prison==0.2.1
protobuf==4.25.0
psutil==5.9.6
pyarrow==13.0.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyee==11.0.1
pyflakes==3.1.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-base-url==2.0.0
pytest-mock==3.12.0
pytest-playwright==0.4.3
pytest-shutil==1.7.0
python-daemon==3.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytzdata==2020.1
querystring-parser==1.2.4
regex==2022.3.2
requests==2.31.0
requests-toolbelt==1.0.0
rich-argparse==1.4.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.34.0
setproctitle==1.3.3
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.32.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.1
tortoise-orm==0.20.0
transformers==4.35.0
Tree==0.2.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
virtualenv==20.25.0
wandb==0.15.12
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
yarg==0.1.9
zstandard==0.22.0
